+++ HT_1.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HT_1", "HT_1\HT_1.csproj", "{08C2D8DC-B259-4346-9E85-3FBA1D27FADE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HT_1", "HT_1\HT_1.csproj", "{08C2D8DC-B259-4346-9E85-3FBA1D27FADE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
